I hate timezones

+++ src/main/kotlin/com/scrumdapp/inviteservice/invites/dto/GroupInviteResponse.kt
@@ -2,7 +2,7 @@ package com.scrumdapp.inviteservice.invites.dto
 
 data class GroupInviteResponse(
     val id: Int,
-    val groupId: Int,
+    val groupId: Long,
     val token: String,
     val expiresAt: String,
     val isActive: Boolean
